backend/requirements.txt
# Django core
Django>=5.2,<6.0
djangorestframework==3.16.1
djangorestframework-simplejwt  # JWT authentication

# Database
psycopg2-binary>=2.9.10,<3.0  # PostgreSQL driver

# Environment variables
python-dotenv>=1.0.0,<2.0

# Password hashing
bcrypt>=4.0.1,<5.0

# CORS for frontend-backend communication
django-cors-headers>=4.0.0,<5.0

# Email sending
django-anymail>=8.4.0,<9.0  # Optional if using SendGrid, Mailgun, etc.

# Asynchronous tasks
celery>=5.6.0,<6.0
redis>=5.3.0,<6.0  # Celery broker

# Payment processing
stripe>=11.0.0,<12.0
paypalrestsdk

# Development tools
ipython>=8.0.0,<9.0
pytest>=8.0.0,<9.0
pytest-django>=4.5.2,<5.0
black>=23.0.0,<24.0
flake8>=6.0.0,<7.0

# Optional utilities
requests>=2.31.0,<3.0
